moving nws23 to package object
